+++ atlasdb-client/src/test/java/com/palantir/atlasdb/logging/SafeLoggableDataUtilsTest.java
@@ -43,13 +43,13 @@ public class SafeLoggableDataUtilsTest {
     private static final String COLUMN_LONG_NAME = "barrrr";
     private static final NameMetadataDescription NAME_METADATA_DESCRIPTION = NameMetadataDescription.create(
             ImmutableList.of(
-                    new NameComponentDescription(
-                            ROW_COMPONENT_NAME,
-                            ValueType.VAR_LONG,
-                            TableMetadataPersistence.ValueByteOrder.ASCENDING,
-                            new UniformRowNamePartitioner(ValueType.VAR_LONG),
-                            null,
-                            LogSafety.SAFE)),
+                    new NameComponentDescription.Builder()
+                            .componentName(ROW_COMPONENT_NAME)
+                            .type(ValueType.VAR_LONG)
+                            .byteOrder(TableMetadataPersistence.ValueByteOrder.ASCENDING)
+                            .uniformRowNamePartitioner(new UniformRowNamePartitioner(ValueType.VAR_LONG))
+                            .logSafety(LogSafety.SAFE)
+                            .build()),
             false);
     private static final ColumnMetadataDescription COLUMN_METADATA_DESCRIPTION =
             new ColumnMetadataDescription(ImmutableList.of(new NamedColumnDescription("bar", COLUMN_LONG_NAME,
@@ -57,7 +57,9 @@ public class SafeLoggableDataUtilsTest {
 
     private static final TableMetadata TABLE_METADATA_1 = new TableMetadata(
             NameMetadataDescription.create(
-                    ImmutableList.of(new NameComponentDescription(ROW_COMPONENT_NAME, ValueType.VAR_LONG)), false),
+                    ImmutableList.of(new NameComponentDescription.Builder()
+                            .componentName(ROW_COMPONENT_NAME).type(ValueType.VAR_LONG).build()),
+                    false),
             new ColumnMetadataDescription(
                     ImmutableList.of(new NamedColumnDescription("bar", "barrrr",
                             ColumnValueDescription.forType(ValueType.VAR_LONG)))),