+++ atlasdb-client/src/test/java/com/palantir/atlasdb/table/description/NameComponentDescriptionTest.java
@@ -20,6 +20,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.Test;
 
+import com.google.common.collect.ImmutableSet;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence;
 import com.palantir.atlasdb.protos.generated.TableMetadataPersistence.LogSafety;
 
@@ -30,20 +31,91 @@ public class NameComponentDescriptionTest {
             TableMetadataPersistence.ValueByteOrder.ASCENDING;
     private static final UniformRowNamePartitioner UNIFORM_ROW_NAME_PARTITIONER =
             new UniformRowNamePartitioner(VALUE_TYPE);
+    private static final ExplicitRowNamePartitioner EXPLICIT_ROW_NAME_PARTITIONER =
+            new ExplicitRowNamePartitioner(VALUE_TYPE, ImmutableSet.of());
 
-    private static final NameComponentDescription DEFAULT_UNNAMED_DESCRIPTION = new NameComponentDescription();
-    private static final NameComponentDescription LOGGABILITY_UNSPECIFIED_DESCRIPTION =
-            new NameComponentDescription(
-                    COMPONENT_NAME,
-                    VALUE_TYPE,
-                    VALUE_BYTE_ORDER,
-                    UNIFORM_ROW_NAME_PARTITIONER,
-                    null);
+    private static final NameComponentDescription DEFAULT_UNNAMED_DESCRIPTION = new NameComponentDescription.Builder()
+            .componentName("name")
+            .type(ValueType.BLOB)
+            .build();
+    private static final NameComponentDescription LOGGABILITY_UNSPECIFIED_DESCRIPTION
+            = new NameComponentDescription.Builder()
+                    .componentName(COMPONENT_NAME)
+                    .type(VALUE_TYPE)
+                    .byteOrder(VALUE_BYTE_ORDER)
+                    .uniformRowNamePartitioner(UNIFORM_ROW_NAME_PARTITIONER)
+                    .explicitRowNamePartitioner(null)
+                    .build();
 
     private static final NameComponentDescription NAME_LOGGABLE_DESCRIPTION =
             createWithSpecifiedLogSafety(LogSafety.SAFE);
     private static final NameComponentDescription NAME_NOT_LOGGABLE_DESCRIPTION =
             createWithSpecifiedLogSafety(LogSafety.UNSAFE);
+
+    @Test
+    public void builderCanCreateNameComponentDescription() {
+        NameComponentDescription description = new NameComponentDescription.Builder()
+                .componentName(COMPONENT_NAME)
+                .type(VALUE_TYPE)
+                .byteOrder(VALUE_BYTE_ORDER)
+                .uniformRowNamePartitioner(UNIFORM_ROW_NAME_PARTITIONER)
+                .explicitRowNamePartitioner(EXPLICIT_ROW_NAME_PARTITIONER)
+                .logSafety(LogSafety.SAFE)
+                .build();
+
+        assertThat(description.getComponentName()).isEqualTo(COMPONENT_NAME);
+        assertThat(description.getType()).isEqualTo(VALUE_TYPE);
+        assertThat(description.getOrder()).isEqualTo(VALUE_BYTE_ORDER);
+        assertThat(description.uniformPartitioner).isEqualTo(UNIFORM_ROW_NAME_PARTITIONER);
+        assertThat(description.getExplicitPartitioner()).isEqualTo(EXPLICIT_ROW_NAME_PARTITIONER);
+        assertThat(description.getLogSafety()).isEqualTo(LogSafety.SAFE);
+    }
+
+    @Test(expected = NullPointerException.class)
+    public void builderRequiresComponentName() {
+        new NameComponentDescription.Builder()
+                .type(VALUE_TYPE)
+                .byteOrder(VALUE_BYTE_ORDER)
+                .uniformRowNamePartitioner(UNIFORM_ROW_NAME_PARTITIONER)
+                .explicitRowNamePartitioner(EXPLICIT_ROW_NAME_PARTITIONER)
+                .logSafety(LogSafety.SAFE)
+                .build();
+    }
+
+    @Test(expected = NullPointerException.class)
+    public void builderRequiresType() {
+        new NameComponentDescription.Builder()
+                .componentName(COMPONENT_NAME)
+                .byteOrder(VALUE_BYTE_ORDER)
+                .uniformRowNamePartitioner(UNIFORM_ROW_NAME_PARTITIONER)
+                .explicitRowNamePartitioner(EXPLICIT_ROW_NAME_PARTITIONER)
+                .logSafety(LogSafety.SAFE)
+                .build();
+    }
+
+    @Test
+    public void builderSetsSaneDefaults() {
+        NameComponentDescription description = new NameComponentDescription.Builder()
+                .componentName(COMPONENT_NAME)
+                .type(VALUE_TYPE)
+                .build();
+
+        assertThat(description.getOrder()).isEqualTo(TableMetadataPersistence.ValueByteOrder.ASCENDING);
+        assertThat(description.uniformPartitioner).isEqualTo(new UniformRowNamePartitioner(VALUE_TYPE));
+        assertThat(description.explicitPartitioner).isNull();
+        assertThat(description.getLogSafety()).isEqualTo(LogSafety.UNSAFE);
+    }
+
+    @Test
+    public void builderCanSetUniformPartitionerToNull() {
+        NameComponentDescription description = new NameComponentDescription.Builder()
+                .componentName(COMPONENT_NAME)
+                .type(VALUE_TYPE)
+                .uniformRowNamePartitioner(null)
+                .build();
+
+        assertThat(description.uniformPartitioner).isNull();
+    }
 
     @Test
     public void nameIsNotLoggableInDefaultDescription() {
@@ -106,12 +178,12 @@ public class NameComponentDescriptionTest {
     }
 
     private static NameComponentDescription createWithSpecifiedLogSafety(LogSafety logSafety) {
-        return new NameComponentDescription(
-                COMPONENT_NAME,
-                VALUE_TYPE,
-                VALUE_BYTE_ORDER,
-                UNIFORM_ROW_NAME_PARTITIONER,
-                null,
-                logSafety);
+        return new NameComponentDescription.Builder()
+                .componentName(COMPONENT_NAME)
+                .type(VALUE_TYPE)
+                .byteOrder(VALUE_BYTE_ORDER)
+                .uniformRowNamePartitioner(UNIFORM_ROW_NAME_PARTITIONER)
+                .logSafety(logSafety)
+                .build();
     }
 }